updates tests to reflect new module
